+++ src/Auth/Implement/JwtWithJsonAuthentication.php
@@ -30,7 +30,9 @@ class JwtWithJsonAuthentication extends \YdbPlatform\Ydb\Auth\JwtAuth
 
     public function getTokenInfo(): TokenInfo
     {
+        $this->logger->debug("YDB: Generating new JWT token");
         $jwt_token = $this->getJwtToken();
+        $this->logger->debug("YDB: Generated new JWT token");
         $request_data = [
             'jwt' => $jwt_token,
         ];