+++ src/Logger/LoggerInterface.php
@@ -5,6 +5,5 @@ namespace YdbPlatform\Ydb\Logger;
 /**
  * @property string $prefix = ""
  */
-interface LoggerInterface extends \Psr\Log\LoggerInterface
-{
+interface LoggerInterface extends \Psr\Log\LoggerInterface{
 }